extract some utils
